Delete onboarding.md
Not ready to be open source yet, we're relying too much on our private Notion at this point. It's not really that the content is confidential (it should be pretty standard), it's just that there will be a lot of links to Notiin that are private (HR processes for instance).